+++ src/controllers/user.controller.js
@@ -2,7 +2,7 @@
 import User from '../entity/user.entity.js';
 import { AppDataSource } from '../config/configDb.js';
 import { userBodyValidation } from '../validations/user.validation.js';
-import { createUserService, getUserService } from '../services/user.service.js';
+import { createUserService, deleteUserService, getUserService, updateUserService } from '../services/user.service.js';
 
 
 export async function createUser(req, res) {
@@ -73,14 +73,11 @@ export async function getUsers(req, res) {
 
 export async function updateUser(req, res) {
     try {
-        const userRepository = AppDataSource.getRepository(User);
-
+        
         const id = req.params.id;
         const user = req.body;
-        
-        const userFound = await userRepository.findOne({
-            where: {id}
-        });
+
+        const userFound = updateUserService(id, user);
 
         if(!userFound) {
             return res.status(404).json({
@@ -88,12 +85,6 @@ export async function updateUser(req, res) {
                 data: null
             });
         }
-
-        await userRepository.update(id, user);
-
-        const userData = await userRepository.findOne({
-            where: {id}
-        });
 
         res.status(200).json({
             message: "Usuario actualizado correctamente",
@@ -107,13 +98,10 @@ export async function updateUser(req, res) {
 
 export async function deleteUser(req, res) {
     try {
-        const userRepository = AppDataSource.getRepository(User);
-
+        
         const id = req.params.id;
 
-        const userFound = await userRepository.findOne({
-            where: {id}
-        });
+        const userFound = deleteUserService(id);
 
         if(!userFound) {
             return res.status(404).json({
@@ -121,8 +109,6 @@ export async function deleteUser(req, res) {
                 data: null
             });
         }
-
-        const userDeleted = await userRepository.remove(userFound);
 
         res.status(200).json({
             message: "Usuario eliminado correctamente",
