+++ src/services/user.service.js
@@ -37,3 +37,48 @@ export async function getUserService(id) {
         console.error("Error al obtener el usuario:", error);
     }
 }
+
+export async function updateUserService(id, user) {
+    try {
+        
+        const userRepository = AppDataSource.getRepository(User);
+        const userFound = await userRepository.findOne({
+            where: {id}
+        });
+
+        if (!userFound) {
+            return null;
+        }
+
+        await userRepository.update(id, user);
+
+        const userData = await userRepository.findOne({
+            where: {id}
+        });
+
+        return userData;
+
+    } catch (error) {
+        console.error("Error al actualizar el usuario:", error);
+    }
+}
+
+export async function deleteUserService(id) {
+
+    try {
+        
+        const userRepository = AppDataSource.getRepository(User);
+
+        const userFound = await userRepository.findOne({
+            where: {id}
+        });
+
+        const userDeleted = await userRepository.remove(userFound);
+
+        return userDeleted;
+
+    } catch (error) {
+        console.error("Error al eliminar el usuario:", error);
+    }
+    
+}
